+++ src/leoStates.ts
@@ -63,6 +63,11 @@ export class LeoStates {
         this._leoIntegration.setTreeViewTitle(
             p_value ? Constants.GUI.TREEVIEW_TITLE : Constants.GUI.TREEVIEW_TITLE_INTEGRATION
         );
+        if (p_value) {
+            this._leoIntegration.serverOpenedFileName = this._leoOpenedFileName; // will trigger outline pane title decorations
+        } else {
+            this._leoIntegration.serverOpenedFileName = "";
+        }
     }
 
     // * Currently opened Leo file path and name, empty string if new unsaved file.
@@ -78,6 +83,7 @@ export class LeoStates {
             this._leoOpenedFileName = "";
             this.qLastContextChange = utils.setContext(Constants.CONTEXT_FLAGS.TREE_TITLED, false);
         }
+        this._leoIntegration.serverOpenedFileName = p_name; // will trigger outline pane title decorations
     }
 
     // * 'states' flags for currently opened tree view
@@ -91,6 +97,7 @@ export class LeoStates {
             this._leoIntegration.refreshDocumentsPane();
         }
         this._leoChanged = p_value;
+        this._leoIntegration.setTreeViewTitle(); // Will refresh the '*' to signify document changed.
         this.qLastContextChange = utils.setContext(Constants.CONTEXT_FLAGS.LEO_CHANGED, p_value);
     }